// continental-frontend/src/interfaces/Api.interface.ts
export enum EmpleadoEstado {
  ASIGNADO = 'Asignado',
  RESERVADO = 'Reservado',
  COMPLETADO = 'Completado',
  TRANSFERIDO = 'Transferido',
  NO_RESPONDIO = 'NoRespondio'
}

// Interfaces para obtener bloques de reservación
export interface EmpleadoBloque {
  empleadoId: number;
  nombreCompleto: string;
  nomina: string;
  posicionEnBloque: number;
  fechaIngreso: string;
  antiguedadAnios: number;
  estado: EmpleadoEstado;
}



export interface BloqueReservacion {
  nombreArea: string;
  duracionHoras: number;
  empleadosAsignados: EmpleadoBloque[];
  id: number;
  grupoId: number;
  nombreGrupo: string;
  numeroBloque: number;
  fechaHoraInicio: string;
  fechaHoraFin: string;
  personasPorBloque: number;
  esBloqueCola: boolean;
  estado: string;
  espaciosDisponibles: number;
}